+++ assets/js/main.js
@@ -1,2 +1,2 @@
-!function(){"use strict";const e=document.getElementById("js-app");let n;function t(e,n){const t=document.createElement(e);return t.classList.add(n),t}function a(e){return Math.floor(100*e)}!async function(){try{n=await fetch("https://blockchain.info/ticker"),n=await n.json()}catch(e){console.log("No data")}finally{!function(){function s(e,n,t){this.currencyName=n,this.unitPrice=e/1e8,this.unitPercentage=a(this.unitPrice),this.smallUnitName=t,this.smallestUnitPrice=this.currencyName===this.smallUnitName?this.unitPrice:100*e/1e8,this.smallestUnitPercentage=a(this.smallestUnitPrice),this.smallestUnitKilled=this.currencyName!=this.smallUnitName&&this.smallestUnitPercentage>100}function c(e,n,a,s,c){const i=t("li","c-currencyitem"),r=t("div","c-currencyitem__main"),l=t("span","c-currencyitem__text");let b,u;r.append(l);const g=e=>{const n=e>100?100:e,a=t("span","c-currencyitem__percentage-bar");return a.style.width=n+"%",a};s>100&&i.classList.add("c-currencyitem--complete"),c?(i.classList.add("c-currencyitem--smallest-complete"),b=`${e} ${a}%`,u=`${n} ${s}%`,l.innerText=u,i.addEventListener("mouseenter",(()=>{l.innerText=b})),i.addEventListener("mouseleave",(()=>{l.innerText=u})),i.append(r),r.append(g(s))):(b=n<1?`${s} <1%`:`${n} ${s}%`,l.innerText=b,i.append(r),r.append(g(s))),o.append(i)}function i(){n=Object.entries(n),n.sort(((e,n)=>e[1].smallestUnitPercentage-n[1].smallestUnitPercentage)),n=Object.fromEntries(n)}function r(e,n,t){switch(e){case"ARS":{const e=new s(n,"Argentine Peso","Argentine Peso");Object.assign(t,e);break}case"AUD":{const e=new s(n,"Australian Dollar","Australian Cent");Object.assign(t,e);break}case"BRL":{const e=new s(n,"Brazilian Real","Brazilian Centavo");Object.assign(t,e);break}case"CAD":{const e=new s(n,"Canadian Dollar","Canadian Cent");Object.assign(t,e);break}case"CHF":{const e=new s(n,"Swiss Franc","Swiss Rappen");Object.assign(t,e);break}case"CLP":{const e=new s(n,"Chilean Peso","Chilean Peso");Object.assign(t,e);break}case"CNY":{const e=new s(n,"Chinese Yuan","Chinese Fen");Object.assign(t,e);break}case"CZK":{const e=new s(n,"Czech Koruna","Czech Haléř");Object.assign(t,e);break}case"DKK":{const e=new s(n,"Danish Krone","Danish Øre");Object.assign(t,e);break}case"EUR":{const e=new s(n,"Euro","Euro Cent");Object.assign(t,e);break}case"GBP":{const e=new s(n,"Pound Sterling","British Penny");Object.assign(t,e);break}case"HKD":{const e=new s(n,"Hong Kong Dollar","Hong Kong Cent");Object.assign(t,e);break}case"HRK":{const e=new s(n,"Croatian Kuna","Croatian Lipa");Object.assign(t,e);break}case"HUF":{const e=new s(n,"Hungarian Forint","Hungarian Forint");Object.assign(t,e);break}case"INR":{const e=new s(n,"Indian Rupee","Indian Paisa");Object.assign(t,e);break}case"ISK":{const e=new s(n,"Icelandic Króna","Icelandic Króna");Object.assign(t,e);break}case"JPY":{const e=new s(n,"Japanese Yen","Japanese Yen");Object.assign(t,e);break}case"KRW":{const e=new s(n,"South Korean Won","South Korean Won");Object.assign(t,e);break}case"NZD":{const e=new s(n,"New Zealand Dollar","New Zealand Cent");Object.assign(t,e);break}case"PLN":{const e=new s(n,"Polish Złoty","Polish Grosz");Object.assign(t,e);break}case"RON":{const e=new s(n,"Romanian Leu","Romanian Ban");Object.assign(t,e);break}case"RUB":{const e=new s(n,"Russian Ruble","Russian Kopek");Object.assign(t,e);break}case"SEK":{const e=new s(n,"Swedish Krona","Swedish Öre");Object.assign(t,e);break}case"SGD":{const e=new s(n,"Singapore Dollar","Singapore Cent");Object.assign(t,e);break}case"THB":{const e=new s(n,"Thai Baht","Thai Satang");Object.assign(t,e);break}case"TRY":{const e=new s(n,"Turkish Lira","Turkish Kuruş");Object.assign(t,e);break}case"TWD":{const e=new s(n,"New Taiwan Dollar","Taiwanese Cent");Object.assign(t,e);break}case"USD":{const e=new s(n,"United States Dollar","United States Cent");Object.assign(t,e);break}}}const o=t("ul","o-currencygrid");e.append(o);for(const e in n){let t=n[e];const a=t.symbol,s=t.sell;["15m","buy","last"].forEach((e=>{delete t[e]})),r(a,s,t)}i();for(const e in n){let t=n[e];c(t.currencyName,t.smallUnitName,t.unitPercentage,t.smallestUnitPercentage,t.smallestUnitKilled)}}()}}(),function(){const e="npub1y73ajm09j3wra8jj9e3h8rkj3xculh520m9plgc6j57rkg7g0yyqt5p5dt",n=`${e.substring(0,15)}...${e.slice(-15)}`,a=t("span","c-nostr-copy__message");a.innerHTML="Copied";const s=document.getElementById("js-nostr-key"),c=document.getElementById("js-nostr-copy"),i=document.querySelector(".js-copy-icon");s.innerText=n,c.addEventListener("click",(()=>{navigator.clipboard.writeText(e).then((function(){console.log("copied"),c.append(a),i.remove(),console.log("copied"),setTimeout((()=>{c.append(i),a.remove()}),2e3)})).catch((function(e){console.error("Failed to copy text to clipboard: ",e)}))}))}()}();
+!function(){"use strict";const e={buildElement:(e,n)=>{const t=document.createElement(e);return t.classList.add(n),t},getPercentage:e=>Math.floor(100*e)};function n(n,t,a){this.currencyName=t,this.unitPrice=n/1e8,this.unitPercentage=e.getPercentage(this.unitPrice),this.smallUnitName=a,this.smallestUnitPrice=this.currencyName===this.smallUnitName?this.unitPrice:100*n/1e8,this.smallestUnitPercentage=e.getPercentage(this.smallestUnitPrice),this.smallestUnitKilled=this.currencyName!=this.smallUnitName&&this.smallestUnitPercentage>100}function t(e,t,a){switch(e){case"ARS":{const e=new n(t,"Argentine Peso","Argentine Peso");Object.assign(a,e);break}case"AUD":{const e=new n(t,"Australian Dollar","Australian Cent");Object.assign(a,e);break}case"BRL":{const e=new n(t,"Brazilian Real","Brazilian Centavo");Object.assign(a,e);break}case"CAD":{const e=new n(t,"Canadian Dollar","Canadian Cent");Object.assign(a,e);break}case"CHF":{const e=new n(t,"Swiss Franc","Swiss Rappen");Object.assign(a,e);break}case"CLP":{const e=new n(t,"Chilean Peso","Chilean Peso");Object.assign(a,e);break}case"CNY":{const e=new n(t,"Chinese Yuan","Chinese Fen");Object.assign(a,e);break}case"CZK":{const e=new n(t,"Czech Koruna","Czech Haléř");Object.assign(a,e);break}case"DKK":{const e=new n(t,"Danish Krone","Danish Øre");Object.assign(a,e);break}case"EUR":{const e=new n(t,"Euro","Euro Cent");Object.assign(a,e);break}case"GBP":{const e=new n(t,"Pound Sterling","British Penny");Object.assign(a,e);break}case"HKD":{const e=new n(t,"Hong Kong Dollar","Hong Kong Cent");Object.assign(a,e);break}case"HRK":{const e=new n(t,"Croatian Kuna","Croatian Lipa");Object.assign(a,e);break}case"HUF":{const e=new n(t,"Hungarian Forint","Hungarian Forint");Object.assign(a,e);break}case"INR":{const e=new n(t,"Indian Rupee","Indian Paisa");Object.assign(a,e);break}case"ISK":{const e=new n(t,"Icelandic Króna","Icelandic Króna");Object.assign(a,e);break}case"JPY":{const e=new n(t,"Japanese Yen","Japanese Yen");Object.assign(a,e);break}case"KRW":{const e=new n(t,"South Korean Won","South Korean Won");Object.assign(a,e);break}case"NZD":{const e=new n(t,"New Zealand Dollar","New Zealand Cent");Object.assign(a,e);break}case"PLN":{const e=new n(t,"Polish Złoty","Polish Grosz");Object.assign(a,e);break}case"RON":{const e=new n(t,"Romanian Leu","Romanian Ban");Object.assign(a,e);break}case"RUB":{const e=new n(t,"Russian Ruble","Russian Kopek");Object.assign(a,e);break}case"SEK":{const e=new n(t,"Swedish Krona","Swedish Öre");Object.assign(a,e);break}case"SGD":{const e=new n(t,"Singapore Dollar","Singapore Cent");Object.assign(a,e);break}case"THB":{const e=new n(t,"Thai Baht","Thai Satang");Object.assign(a,e);break}case"TRY":{const e=new n(t,"Turkish Lira","Turkish Kuruş");Object.assign(a,e);break}case"TWD":{const e=new n(t,"New Taiwan Dollar","Taiwanese Cent");Object.assign(a,e);break}case"USD":{const e=new n(t,"United States Dollar","United States Cent");Object.assign(a,e);break}}}const a=document.getElementById("js-app");let s;!async function(){try{s=await fetch("https://blockchain.info/ticker"),s=await s.json()}catch(e){console.log("No data")}finally{!function(){function n(){for(const e in s){let n=s[e];const a=n.symbol,c=n.sell;["15m","buy","last"].forEach((e=>{delete n[e]})),t(a,c,n)}}function c(){s=Object.entries(s),s.sort(((e,n)=>e[1].smallestUnitPercentage-n[1].smallestUnitPercentage)),s=Object.fromEntries(s)}n(),c(),function(){const n=e.buildElement("ul","o-currencygrid");for(const e in s){let n=s[e];t(n.currencyName,n.smallUnitName,n.unitPercentage,n.smallestUnitPercentage,n.smallestUnitKilled)}function t(t,a,s,c,i){const r=e.buildElement("li","c-currencyitem"),o=e.buildElement("div","c-currencyitem__main"),l=e.buildElement("span","c-currencyitem__text");let b,u;o.append(l);const g=n=>{const t=n>100?100:n,a=e.buildElement("span","c-currencyitem__percentage-bar");return a.style.width=t+"%",a};c>100&&r.classList.add("c-currencyitem--complete"),i?(r.classList.add("c-currencyitem--smallest-complete"),b=`${t} ${s}%`,u=`${a} ${c}%`,l.innerText=u,r.addEventListener("mouseenter",(()=>{l.innerText=b})),r.addEventListener("mouseleave",(()=>{l.innerText=u})),r.append(o),o.append(g(c))):(b=a<1?`${c} <1%`:`${a} ${c}%`,l.innerText=b,r.append(o),o.append(g(c))),n.append(r)}a.append(n)}()}()}}(),function(){const n="npub1y73ajm09j3wra8jj9e3h8rkj3xculh520m9plgc6j57rkg7g0yyqt5p5dt",t=`${n.substring(0,15)}...${n.slice(-15)}`,a=e.buildElement("span","c-nostr-copy__message");a.innerHTML="Copied";const s=document.getElementById("js-nostr-key"),c=document.getElementById("js-nostr-copy"),i=document.querySelector(".js-copy-icon");s.innerText=t,c.addEventListener("click",(()=>{navigator.clipboard.writeText(n).then((function(){console.log("copied"),c.append(a),i.remove(),console.log("copied"),setTimeout((()=>{c.append(i),a.remove()}),2e3)})).catch((function(e){console.error("Failed to copy text to clipboard: ",e)}))}))}()}();
 //# sourceMappingURL=main.js.map
